+++ case_simple_pipe.f08
@@ -38,18 +38,18 @@ contains
         !
         character(64) :: subroutine_name = 'case_simple_pipe_initialize'
 
-        real,  intent(in)  :: pipe_length, pipe_breadth
-        real,  intent(in)  :: pipe_topwidth, subdivide_length, full_depth
-        real,  intent(in)  :: lowerZ, upperZ, ManningsN, initial_flowrate
-        real,  intent(in)  :: depth_upstream, depth_dnstream, init_depth
+        real(8),  intent(in)  :: pipe_length, pipe_breadth
+        real(8),  intent(in)  :: pipe_topwidth, subdivide_length, full_depth
+        real(8),  intent(in)  :: lowerZ, upperZ, ManningsN, initial_flowrate
+        real(8),  intent(in)  :: depth_upstream, depth_dnstream, init_depth
 
         integer, intent(in):: roughness_type, idepth_type
 
         integer,   dimension(:,:), allocatable, target, intent(out)    :: linkI
         integer,   dimension(:,:), allocatable, target, intent(out)    :: nodeI
 
-        real,      dimension(:,:), allocatable, target, intent(out)    :: linkR
-        real,      dimension(:,:), allocatable, target, intent(out)    :: nodeR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: linkR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: nodeR
 
         logical,   dimension(:,:), allocatable, target, intent(out)    :: linkYN
         logical,   dimension(:,:), allocatable, target, intent(out)    :: nodeYN
@@ -115,18 +115,18 @@ contains
         !
         character(64) :: subroutine_name = 'case_simple_pipe_links_and_nodes'
 
-        real,  intent(in)  :: pipe_length, pipe_breadth
-        real,  intent(in)  :: pipe_topwidth, subdivide_length, full_depth
-        real,  intent(in)  :: lowerZ, upperZ, ManningsN, initial_flowrate
-        real,  intent(in)  :: depth_upstream, depth_dnstream, init_depth
+        real(8),  intent(in)  :: pipe_length, pipe_breadth
+        real(8),  intent(in)  :: pipe_topwidth, subdivide_length, full_depth
+        real(8),  intent(in)  :: lowerZ, upperZ, ManningsN, initial_flowrate
+        real(8),  intent(in)  :: depth_upstream, depth_dnstream, init_depth
 
         integer, intent(in):: roughness_type, idepth_type
 
         integer,   dimension(:,:), allocatable, target, intent(out)    :: linkI
         integer,   dimension(:,:), allocatable, target, intent(out)    :: nodeI
 
-        real,      dimension(:,:), allocatable, target, intent(out)    :: linkR
-        real,      dimension(:,:), allocatable, target, intent(out)    :: nodeR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: linkR
+        real(8),      dimension(:,:), allocatable, target, intent(out)    :: nodeR
 
         logical,   dimension(:,:), allocatable, target, intent(out)    :: linkYN
         logical,   dimension(:,:), allocatable, target, intent(out)    :: nodeYN
